Move Nostr code to NostrFeedParser, in it's own package.
